requirements.txt
django-cors-headers==3.5.0
django-heroku==0.3.1
Django==3.0.5
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
